+++ Changelog.md
@@ -10,6 +10,7 @@
 
 ### Fixed
 - Fix the issue where the step size calculated by the `extrapolate()` function in `external_control.script` is a fixed steptime.
+- Fix the error of missing the "float" type when parsing the 30001 robot error exception message.
 
 ## [v1.2.0] - 2025-08-14